project/backend/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.44
pydantic==2.5.0
cryptography==41.0.7
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1